+++ src/private/rapidxml_formatter.cpp
@@ -469,7 +469,7 @@ namespace adm {
       node.addOptionalElement<Height>(&audioBlock, "height");
       node.addOptionalElement<Depth>(&audioBlock, "depth");
       node.addOptionalElement<Cartesian>(&audioBlock, "cartesian");
-      node.addOptionalElement<Gain>(&audioBlock, "gain");
+      node.addOptionalElement<Gain>(&audioBlock, "gain", &formatGain);
       node.addOptionalElement<Diffuse>(&audioBlock, "diffuse");
       node.addOptionalElement<ChannelLock>(&audioBlock, "channelLock", &formatChannelLock);
       node.addOptionalElement<ObjectDivergence>(&audioBlock, "objectDivergence", &formatObjectDivergence);
@@ -504,6 +504,15 @@ namespace adm {
         parentNode.addOptionalElement<Z>(
             &cartesianPosition, name,
             detail::formatMultiElementAttribute("coordinate", "Z"));
+      }
+    }
+
+    void formatGain(XmlNode &node, const Gain &gain) {
+      if (gain.isLinear()) {
+        node.setValue(gain.asLinear());
+      } else {
+        node.setValue(gain.asDb());
+        node.addAttribute("gainUnit", "dB");
       }
     }